doc: update the deprecation for `process.exitCode` to clarify its scope
This updates DEP0164 to clarify its deprecation scope.

`process.exitCode` coercion to integer has been EOL in DEP016.
`process.exitCode` is implemented as a setter as a non-configurable
property to validate the given value.

A user behaviour reported was to do the coercion to `0` by doing
`delete process.exitCode`. As `process.exitCode` is non-configurable,
this behaviour is deprecated as well.

Signed-off-by: Daeyeon Jeong <daeyeon.dev@gmail.com>